requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai-whisper==20231117
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==2.4.2
llama-cpp-python==0.2.56
huggingface_hub==0.20.3
transformers==4.38.2
accelerate==0.27.2
